remove internal implementation of longestCommonSubsequence
Back in 2019, the usage of this method in BsonDiff was changed to use the commons-collections4 ListUtils version.  There are no other usages of this internal, package-visible static method.  So, I'm removing it from the code base.  It's been six years, if we haven't found a use for it in that time, we probably never will.  Removing increases unit test line coverage from 91% to 94% :-)